images/cert-syncer/requirements.txt
pyyaml==6.0.3
valkey==6.1.1
